requirements.txt
Flask
pandas
numpy
matplotlib
pymongo
python-dotenv